services/asr_whisper/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ekho-asr-whisper"
version = "0.1.0"
description = "ASR service using Whisper for EKHO platform"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.17",
    "openai-whisper>=20240930",
    "ekho-core>=0.1.0",
    "torch>=2.5.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "black>=24.0",
    "ruff>=0.6",
]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501", "B008", "I001"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
